Undo the segregation of security sources
With this change we move security profile sources into the main package and
remove the security profile from the POM. As a result HBase will no longer
compile against a Hadoop version 0.20 or earlier without security APIs. When
we introduce crypto it becomes impossible to separate all security changes
to an optional package.