Client side per-cell encryption
This change adds EncryptedHTable and EncryptedHTableFactory. The former
performs transparent client side encryption of values. The latter is for use
with HTablePool to instantiate EncryptedHTable instances as needed. Because
we encrypt value data client side there are limitations: For any family
configured for encryption (by default - all), checkAndPut, checkAndDelete,
Append, Increment, incrementColumnValue, and filters that operate on value
data in a server side context will not work. This is because the server side
context does not have key material for decrypting the data there.